requirements.txt
gunicorn==20.1.0
flask
flask_cors
flask_sqlalchemy
flask_login
flask_caching
sqlalchemy
python-dotenv